backend/requirements.txt
anthropic>=0.49.0
fastapi>=0.115.0
httpx>=0.27.0
pydantic>=2.7.0
pymorphy2>=0.9.1
python-dotenv>=1.0.1
python-multipart>=0.0.9
pytest>=8.2.0
setuptools>=70.0.0,<81.0.0
uvicorn[standard]>=0.30.0
